refactor(cli): move tooling to cli/tools/
