+++ cli/tools/mod.rs
@@ -1,0 +1,8 @@
+// Copyright 2018-2020 the Deno authors. All rights reserved. MIT license.
+
+pub mod fmt;
+pub mod installer;
+pub mod lint;
+pub mod repl;
+pub mod test_runner;
+pub mod upgrade;
